refactor($module): Refactor Hype5 for readability and modularity.
Refactored methods for better composition and lessen the tight coupling. Changes made to improve
readability and increase the ease of adding new method and features in the feature. exports are now
created via factory instead of explicity. I'm going to the stop writing now because this is more
detail then necessary... P.S. Childish Gambino's Because of the Internet is still a flipping
classic... 😎